move db helper
