+++ test/local-integration/services/database-service.test.js
@@ -1,6 +1,6 @@
 describe('database service tests', () => {
   const databaseService = require('../../../server/services/database-service')
-  const dbHelper = require('./db-helper')
+  const dbHelper = require('../../db-helper')
 
   afterAll(async () => {
     await dbHelper.close()
