+++ test/local-integration/services/payment-service.test.js
@@ -1,5 +1,4 @@
-
-const dbHelper = require('./db-helper')
+const dbHelper = require('../../db-helper')
 const paymentService = require('../../../server/services/payment-service')
 
 const payment1 = {
